add bos_clinet

+++ app/models/concerns/upload_util.rb
@@ -5,7 +5,7 @@ class UploadUtil
   URL = "http://upload.topdmc.com.cn"
 
   class << self
-    def upload(category, file)
+    def upload(dir, file)
 
       if file.is_a? File
         path = file.path
@@ -17,18 +17,30 @@ class UploadUtil
         temp.close
         path = temp.path
       end
+      filename = "#{Time.now.to_i}-#{SecureRandom.hex(8)}.xlsx"
+      @bucket = BosClient::Bucket.new(name: Rails.application.secrets.dmc_ro)
+      ret = BosClient::Object.upload(path: dir,
+                                file: path,
+                                filename: filename,
+                                bucket: @bucket)
 
-      res = JSON.parse(RestClient.post("#{URL}?category=#{category}", :file => File.open(path)).body)
-
-      if res.is_a? Array
-        url = res[0]['url']
-      elsif res.is_a?(Hash)
-        err = res['message']
+      if ret
+        err = nil
+        url = "#{@bucket.bucket_host}/#{dir}/#{filename}"
+        temp.unlink if temp.present?
+        return url, err
       end
-
-      temp.unlink if temp.present?
-
-      return url, err
+      #res = JSON.parse(RestClient.post("#{URL}?category=#{category}", :file => File.open(path)).body)
+      #
+      # if res.is_a? Array
+      #   url = res[0]['url']
+      # elsif res.is_a?(Hash)
+      #   err = res['message']
+      # end
+      #
+      # temp.unlink if temp.present?
+      #
+      # return url, err
     end
   end
 end
